Fix wheel builds (#246)
fixes the missing shared library issue that a number of users experienced with pip due to issues with installation refactor. Now behavior in build_py command is overridden to download cspice in that step, which is run during/near the top of all wheel/build runs.  Seems to fix issues with pip wheel and it assuming that the wheels are generic, and tested  locally and with version deployed to testpypi.