added almost all remaining N67 wrappers using wrapper generator script
most functions missing tests, due to complexity in parsing some
parameters are missing from some functions and argtypes/conversions are
probably not correct especially for pointers. Tests still need to be
added for all of these functions. For certain functions that used the
f2c functions directly the test results change when using the new _c
wrapper functions. Need to investigate that...